requirements.txt
arxiv
beautifulsoup4
chromadb
fastapi
faiss-cpu
gradio
h11
huggingface_hub
httpcore
httpx
ipykernel
langchain
langchain-chroma
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-openai
langchain-text-splitters
numexpr
openai
pypdf
python-dotenv
sentence-transformers
sse-starlette
sqlalchemy
streamlit
tqdm
transformers
unstructured
validators
uvicorn
youtube-transcript-api
wikipedia
ddgs
duckduckgo-search
cassio
PyPDF2
gradio